Create config module
